chore: fix agent content overflow issue
