+++ frontend/src/components/Agent/AgentTab/AgentTabItem.tsx
