+++ frontend/src/components/Agent/shared/AgentAvatar.tsx
