+++ frontend/src/components/Agent/shared/ErrorPlaceholder.tsx
@@ -1,4 +1,4 @@
-import { cn } from '../lib/utils';
+import { cn } from '../../lib/utils';
 
 const ErrorPlaceholder = ({
     className,
