+++ frontend/src/components/Agent/shared/GroupParameters.tsx
