+++ frontend/src/components/Modals/ModalViews/AgentManualTriggerModalView.tsx
@@ -6,7 +6,7 @@ import { useMutation } from '@tanstack/react-query';
 import { validateInputFieldForGroup } from '@utils';
 import { useAtom } from 'jotai';
 
-import GroupParams from '@app/components/Agent/GroupParameters';
+import GroupParams from '@app/components/Agent/shared/GroupParameters';
 import { Button } from '@app/components/atoms/Button';
 import { Input } from '@app/components/atoms/Input';
 import { ErrorToast, SuccessToast } from '@app/components/molecules/CustomToasts';
